+++ src/compas_timber/connections/__init__.py
@@ -21,10 +21,15 @@ from .t_dovetail import TDovetailJoint
 from .t_tenon_mortise import TenonMortiseJoint
 from .ball_node import BallNodeJoint
 from .y_butt import YButtJoint
+from .k_butt import KButtJoint
+from .k_miter import KMiterJoint
 from .oligina import OliGinaJoint
 from .utilities import beam_ref_side_incidence
 from .utilities import beam_ref_side_incidence_with_vector
 from .utilities import point_centerline_towards_joint
+from .utilities import extend_main_beam_to_cross_beam
+from .utilities import angle_and_dot_product_main_beam_and_cross_beam
+from .utilities import parse_cross_beam_and_main_beams_from_cluster
 from .plate_joint import PlateJoint
 from .panel_joint import PanelJoint
 from .plate_butt_joint import PlateButtJoint
@@ -41,6 +46,7 @@ from .analyzers import CompositeAnalyzer
 from .analyzers import Cluster
 from .analyzers import BeamGroupAnalyzer
 from .analyzers import MaxNCompositeAnalyzer
+
 
 __all__ = [
     "Joint",
@@ -68,9 +74,14 @@ __all__ = [
     "TenonMortiseJoint",
     "YButtJoint",
     "OliGinaJoint",
+    "KMiterJoint",
+    "KButtJoint",
     "beam_ref_side_incidence",
     "beam_ref_side_incidence_with_vector",
     "point_centerline_towards_joint",
+    "extend_main_beam_to_cross_beam",
+    "angle_and_dot_product_main_beam_and_cross_beam",
+    "parse_cross_beam_and_main_beams_from_cluster",
     "PlateJoint",
     "PanelJoint",
     "PlateButtJoint",